fixed windows build

+++ src/cu_std/win32/tthread.cpp
@@ -3,7 +3,7 @@
 #include "aallocator.h"
 
 struct InternalArgs{
-    s32 (*threadproc)(void*);
+    s64 (*threadproc)(void*);
     void* args;
 };
